+++ test/integration/integration_test.go
@@ -60,37 +60,35 @@ func TestMain(m *testing.M) {
 
 // Deployment ----------------------------------------------------------------------------------------------------------
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted (see https://github.com/kubernetes/kubernetes/pull/130183 /
-//  https://github.com/kubernetes/enhancements/blob/master/keps/sig-node/1287-in-place-update-pod-resources/README.md#memory-limit-decreases).
-//func TestDeploymentNonStartupAdmittedFlowStartupProbeAll(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-non-startup-admitted-flow-startup-probe-all"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsAllDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
-//			config.removeReadinessProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//	)
-//}
+func TestDeploymentNonStartupAdmittedFlowStartupProbeAll(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-non-startup-admitted-flow-startup-probe-all"
+	maybeRegisterCleanup(t, namespace)
+
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsAllDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
+			config.removeReadinessProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+	)
+}
 
 func TestDeploymentNonStartupAdmittedFlowStartupProbeCpu(t *testing.T) {
 	t.Parallel()
@@ -122,67 +120,65 @@ func TestDeploymentNonStartupAdmittedFlowStartupProbeCpu(t *testing.T) {
 	)
 }
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentNonStartupAdmittedFlowStartupProbeMemory(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-non-startup-admitted-flow-startup-probe-memory"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsMemoryOnlyDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
-//			config.removeReadinessProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//	)
-//}
+func TestDeploymentNonStartupAdmittedFlowStartupProbeMemory(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-non-startup-admitted-flow-startup-probe-memory"
+	maybeRegisterCleanup(t, namespace)
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentStartupAdmittedFlowStartupProbeAll(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-startup-admitted-flow-startup-probe-all"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsAllDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
-//			config.removeReadinessProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//	)
-//}
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsMemoryOnlyDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
+			config.removeReadinessProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+	)
+}
+
+func TestDeploymentStartupAdmittedFlowStartupProbeAll(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-startup-admitted-flow-startup-probe-all"
+	maybeRegisterCleanup(t, namespace)
+
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsAllDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
+			config.removeReadinessProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+	)
+}
 
 func TestDeploymentStartupAdmittedFlowStartupProbeCpu(t *testing.T) {
 	t.Parallel()
@@ -214,67 +210,65 @@ func TestDeploymentStartupAdmittedFlowStartupProbeCpu(t *testing.T) {
 	)
 }
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentStartupAdmittedFlowStartupProbeMemory(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-startup-admitted-flow-startup-probe-memory"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsMemoryOnlyDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
-//			config.removeReadinessProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//		},
-//	)
-//}
+func TestDeploymentStartupAdmittedFlowStartupProbeMemory(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-startup-admitted-flow-startup-probe-memory"
+	maybeRegisterCleanup(t, namespace)
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentNonStartupAdmittedFlowReadinessProbeAll(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-non-startup-admitted-flow-readiness-probe-all"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsAllDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
-//			config.removeStartupProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//	)
-//}
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsMemoryOnlyDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
+			config.removeReadinessProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+		},
+	)
+}
+
+func TestDeploymentNonStartupAdmittedFlowReadinessProbeAll(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-non-startup-admitted-flow-readiness-probe-all"
+	maybeRegisterCleanup(t, namespace)
+
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsAllDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
+			config.removeStartupProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+	)
+}
 
 func TestDeploymentNonStartupAdmittedFlowReadinessProbeCpu(t *testing.T) {
 	t.Parallel()
@@ -306,67 +300,65 @@ func TestDeploymentNonStartupAdmittedFlowReadinessProbeCpu(t *testing.T) {
 	)
 }
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentNonStartupAdmittedFlowReadinessProbeMemory(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-non-startup-admitted-flow-readiness-probe-memory"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsMemoryOnlyDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
-//			config.removeStartupProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//	)
-//}
+func TestDeploymentNonStartupAdmittedFlowReadinessProbeMemory(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-non-startup-admitted-flow-readiness-probe-memory"
+	maybeRegisterCleanup(t, namespace)
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentStartupAdmittedFlowReadinessProbeAll(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-startup-admitted-flow-readiness-probe-all"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsAllDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
-//			config.removeStartupProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//	)
-//}
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsMemoryOnlyDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardPostStartup(namespace, int32(replicas), annotations)
+			config.removeStartupProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+	)
+}
+
+func TestDeploymentStartupAdmittedFlowReadinessProbeAll(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-startup-admitted-flow-readiness-probe-all"
+	maybeRegisterCleanup(t, namespace)
+
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsAllDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
+			config.removeStartupProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+	)
+}
 
 func TestDeploymentStartupAdmittedFlowReadinessProbeCpu(t *testing.T) {
 	t.Parallel()
@@ -398,89 +390,87 @@ func TestDeploymentStartupAdmittedFlowReadinessProbeCpu(t *testing.T) {
 	)
 }
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentStartupAdmittedFlowReadinessProbeMemory(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-startup-admitted-flow-readiness-probe-memory"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	testWorkflow(
-//		t,
-//		namespace,
-//		csaQuantityAnnotationsMemoryOnlyDefault,
-//		func(annotations csaQuantityAnnotations) (string, int) {
-//			replicas := 2
-//			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
-//			config.removeStartupProbes()
-//			return config.deploymentJson(), replicas
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, false)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
-//		},
-//		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]pod.StatusAnnotation) {
-//			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
-//		},
-//	)
-//}
+func TestDeploymentStartupAdmittedFlowReadinessProbeMemory(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-startup-admitted-flow-readiness-probe-memory"
+	maybeRegisterCleanup(t, namespace)
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentScaleWhenUnknownResourcesAll(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-scale-when-unknown-resources-all"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	_ = kubeDeleteNamespace(t, namespace)
-//	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
-//
-//	annotations := csaQuantityAnnotations{
-//		cpuStartup:                "150m",
-//		cpuPostStartupRequests:    "100m",
-//		cpuPostStartupLimits:      "100m",
-//		memoryStartup:             "150M",
-//		memoryPostStartupRequests: "100M",
-//		memoryPostStartupLimits:   "100M",
-//	}
-//
-//	config := echoDeploymentConfigStandard(
-//		namespace,
-//		2,
-//		annotations,
-//		"175m", "175m",
-//		"175M", "175M",
-//		echoServerDefaultProbeInitialDelaySeconds,
-//	)
-//	config.removeReadinessProbes()
-//	maybeLogErrAndFailNow(t, kubeApplyYamlOrJsonResources(t, config.deploymentJson()))
-//
-//	names, err := kubeGetPodNames(t, namespace, echoServerName)
-//	maybeLogErrAndFailNow(t, err)
-//
-//	podStatusAnn, errs := csaWaitStatusAll(t, namespace, names, csaStatusMessageStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
-//	if len(errs) > 0 {
-//		maybeLogErrAndFailNow(t, errs[len(errs)-1])
-//	}
-//
-//	assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//
-//	podStatusAnn, errs = csaWaitStatusAll(t, namespace, names, csaStatusMessagePostStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
-//	if len(errs) > 0 {
-//		maybeLogErrAndFailNow(t, errs[len(errs)-1])
-//	}
-//
-//	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupCommandedUnknownRes, namespace, names)
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupEnacted, namespace, names)
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupCommanded, namespace, names)
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupEnacted, namespace, names)
-//	assertEventCount(t, 4, namespace, names)
-//}
+	testWorkflow(
+		t,
+		namespace,
+		csaQuantityAnnotationsMemoryOnlyDefault,
+		func(annotations csaQuantityAnnotations) (string, int) {
+			replicas := 2
+			config := echoDeploymentConfigStandardStartup(namespace, int32(replicas), annotations)
+			config.removeStartupProbes()
+			return config.deploymentJson(), replicas
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, false)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertStartupEnacted(t, annotations, podStatusAnn, false, true, true)
+		},
+		func(t *testing.T, annotations csaQuantityAnnotations, podStatusAnn map[*v1.Pod]podcommon.StatusAnnotation) {
+			assertPostStartupEnacted(t, annotations, podStatusAnn, false, true)
+		},
+	)
+}
+
+func TestDeploymentScaleWhenUnknownResourcesAll(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-scale-when-unknown-resources-all"
+	maybeRegisterCleanup(t, namespace)
+
+	_ = kubeDeleteNamespace(t, namespace)
+	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
+
+	annotations := csaQuantityAnnotations{
+		cpuStartup:                "150m",
+		cpuPostStartupRequests:    "100m",
+		cpuPostStartupLimits:      "100m",
+		memoryStartup:             "150M",
+		memoryPostStartupRequests: "100M",
+		memoryPostStartupLimits:   "100M",
+	}
+
+	config := echoDeploymentConfigStandard(
+		namespace,
+		2,
+		annotations,
+		"175m", "175m",
+		"175M", "175M",
+		echoServerDefaultProbeInitialDelaySeconds,
+	)
+	config.removeReadinessProbes()
+	maybeLogErrAndFailNow(t, kubeApplyYamlOrJsonResources(t, config.deploymentJson()))
+
+	names, err := kubeGetPodNames(t, namespace, echoServerName)
+	maybeLogErrAndFailNow(t, err)
+
+	podStatusAnn, errs := csaWaitStatusAll(t, namespace, names, csaStatusMessageStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
+	if len(errs) > 0 {
+		maybeLogErrAndFailNow(t, errs[len(errs)-1])
+	}
+
+	assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+
+	podStatusAnn, errs = csaWaitStatusAll(t, namespace, names, csaStatusMessagePostStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
+	if len(errs) > 0 {
+		maybeLogErrAndFailNow(t, errs[len(errs)-1])
+	}
+
+	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupCommandedUnknownRes, namespace, names)
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupEnacted, namespace, names)
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupCommanded, namespace, names)
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupEnacted, namespace, names)
+	assertEventCount(t, 4, namespace, names)
+}
 
 func TestDeploymentScaleWhenUnknownResourcesCpu(t *testing.T) {
 	t.Parallel()
@@ -531,68 +521,66 @@ func TestDeploymentScaleWhenUnknownResourcesCpu(t *testing.T) {
 	assertEventCount(t, 4, namespace, names)
 }
 
-// TODO(wt) temporarily disabled until memory downscaling is permitted.
-//func TestDeploymentScaleWhenUnknownResourcesMemory(t *testing.T) {
-//	t.Parallel()
-//	namespace := "deployment-scale-when-unknown-resources-memory"
-//	maybeRegisterCleanup(t, namespace)
-//
-//	_ = kubeDeleteNamespace(t, namespace)
-//	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
-//
-//	annotations := csaQuantityAnnotations{
-//		memoryStartup:             "150M",
-//		memoryPostStartupRequests: "100M",
-//		memoryPostStartupLimits:   "100M",
-//	}
-//
-//	config := echoDeploymentConfigStandard(
-//		namespace,
-//		2,
-//		annotations,
-//		echoServerCpuDisabledRequests, echoServerCpuDisabledLimits,
-//		"175M", "175M",
-//		echoServerDefaultProbeInitialDelaySeconds,
-//	)
-//	config.removeReadinessProbes()
-//	maybeLogErrAndFailNow(t, kubeApplyYamlOrJsonResources(t, config.deploymentJson()))
-//
-//	names, err := kubeGetPodNames(t, namespace, echoServerName)
-//	maybeLogErrAndFailNow(t, err)
-//
-//	podStatusAnn, errs := csaWaitStatusAll(t, namespace, names, csaStatusMessageStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
-//	if len(errs) > 0 {
-//		maybeLogErrAndFailNow(t, errs[len(errs)-1])
-//	}
-//
-//	assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
-//
-//	podStatusAnn, errs = csaWaitStatusAll(t, namespace, names, csaStatusMessagePostStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
-//	if len(errs) > 0 {
-//		maybeLogErrAndFailNow(t, errs[len(errs)-1])
-//	}
-//
-//	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
-//
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupCommandedUnknownRes, namespace, names)
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupEnacted, namespace, names)
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupCommanded, namespace, names)
-//	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupEnacted, namespace, names)
-//	assertEventCount(t, 4, namespace, names)
-//}
+func TestDeploymentScaleWhenUnknownResourcesMemory(t *testing.T) {
+	t.Parallel()
+	namespace := "deployment-scale-when-unknown-resources-memory"
+	maybeRegisterCleanup(t, namespace)
+
+	_ = kubeDeleteNamespace(t, namespace)
+	maybeLogErrAndFailNow(t, kubeCreateNamespace(t, namespace))
+
+	annotations := csaQuantityAnnotations{
+		memoryStartup:             "150M",
+		memoryPostStartupRequests: "100M",
+		memoryPostStartupLimits:   "100M",
+	}
+
+	config := echoDeploymentConfigStandard(
+		namespace,
+		2,
+		annotations,
+		echoServerCpuDisabledRequests, echoServerCpuDisabledLimits,
+		"175M", "175M",
+		echoServerDefaultProbeInitialDelaySeconds,
+	)
+	config.removeReadinessProbes()
+	maybeLogErrAndFailNow(t, kubeApplyYamlOrJsonResources(t, config.deploymentJson()))
+
+	names, err := kubeGetPodNames(t, namespace, echoServerName)
+	maybeLogErrAndFailNow(t, err)
+
+	podStatusAnn, errs := csaWaitStatusAll(t, namespace, names, csaStatusMessageStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
+	if len(errs) > 0 {
+		maybeLogErrAndFailNow(t, errs[len(errs)-1])
+	}
+
+	assertStartupEnacted(t, annotations, podStatusAnn, true, false, true)
+
+	podStatusAnn, errs = csaWaitStatusAll(t, namespace, names, csaStatusMessagePostStartupEnacted, testsDefaultWaitStatusTimeoutSecs)
+	if len(errs) > 0 {
+		maybeLogErrAndFailNow(t, errs[len(errs)-1])
+	}
+
+	assertPostStartupEnacted(t, annotations, podStatusAnn, true, false)
+
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupCommandedUnknownRes, namespace, names)
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessageStartupEnacted, namespace, names)
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupCommanded, namespace, names)
+	assertEvent(t, kubeEventTypeNormal, csaEventReasonScaling, csaStatusMessagePostStartupEnacted, namespace, names)
+	assertEventCount(t, 4, namespace, names)
+}
 
 // StatefulSet ---------------------------------------------------------------------------------------------------------
 
-// TODO(wt) temporarily changed from TestStatefulSetFlowStartupProbeAll until memory downscaling is permitted.
-func TestStatefulSetFlowStartupProbeCpu(t *testing.T) {
+func TestStatefulSetFlowStartupProbeAll(t *testing.T) {
 	t.Parallel()
-	namespace := "statefulset-flow-startup-probe-cpu" // TODO(wt) temporarily changed from statefulset-flow-startup-probe-all.
+	namespace := "statefulset-flow-startup-probe-all"
 	maybeRegisterCleanup(t, namespace)
 
 	testWorkflow(
 		t,
 		namespace,
-		csaQuantityAnnotationsCpuOnlyDefault, // TODO(wt) temporarily changed from csaQuantityAnnotationsAllDefault.
+		csaQuantityAnnotationsAllDefault,
 		func(annotations csaQuantityAnnotations) (string, int) {
 			replicas := 1 // Can only test with 1 replica since pods are started sequentially (after become ready).
 			config := echoStatefulSetConfigStandardPostStartup(namespace, int32(replicas), annotations)
@@ -614,16 +602,15 @@ func TestStatefulSetFlowStartupProbeCpu(t *testing.T) {
 	)
 }
 
-// TODO(wt) temporarily changed from TestStatefulSetFlowReadinessProbeAll until memory downscaling is permitted.
-func TestStatefulSetFlowReadinessProbeCpu(t *testing.T) {
+func TestStatefulSetFlowReadinessProbeAll(t *testing.T) {
 	t.Parallel()
-	namespace := "statefulset-flow-readiness-probe-cpu" // TODO(wt) temporarily changed from statefulset-flow-readiness-probe-all.
+	namespace := "statefulset-flow-readiness-probe-all"
 	maybeRegisterCleanup(t, namespace)
 
 	testWorkflow(
 		t,
 		namespace,
-		csaQuantityAnnotationsCpuOnlyDefault, // TODO(wt) temporarily changed from csaQuantityAnnotationsAllDefault.
+		csaQuantityAnnotationsAllDefault,
 		func(annotations csaQuantityAnnotations) (string, int) {
 			replicas := 1
 			config := echoStatefulSetConfigStandardPostStartup(namespace, int32(replicas), annotations)
@@ -647,16 +634,15 @@ func TestStatefulSetFlowReadinessProbeCpu(t *testing.T) {
 
 // DaemonSet -----------------------------------------------------------------------------------------------------------
 
-// TODO(wt) temporarily changed from TestDaemonSetFlowStartupProbeAll until memory downscaling is permitted.
-func TestDaemonSetFlowStartupProbeCpu(t *testing.T) {
+func TestDaemonSetFlowStartupProbeAll(t *testing.T) {
 	t.Parallel()
-	namespace := "daemonset-flow-startup-probe-cpu" // TODO(wt) temporarily changed from daemonset-flow-startup-probe-all.
+	namespace := "daemonset-flow-startup-probe-all"
 	maybeRegisterCleanup(t, namespace)
 
 	testWorkflow(
 		t,
 		namespace,
-		csaQuantityAnnotationsCpuOnlyDefault, // TODO(wt) temporarily changed from csaQuantityAnnotationsAllDefault.
+		csaQuantityAnnotationsAllDefault,
 		func(annotations csaQuantityAnnotations) (string, int) {
 			config := echoDaemonSetConfigStandardPostStartup(namespace, annotations)
 			config.removeReadinessProbes()
@@ -677,16 +663,15 @@ func TestDaemonSetFlowStartupProbeCpu(t *testing.T) {
 	)
 }
 
-// TODO(wt) temporarily changed from TestDaemonSetFlowReadinessProbeAll until memory downscaling is permitted.
-func TestDaemonSetFlowReadinessProbeCpu(t *testing.T) {
+func TestDaemonSetFlowReadinessProbeAll(t *testing.T) {
 	t.Parallel()
-	namespace := "daemonset-flow-readiness-probe-cpu" // TODO(wt) temporarily changed from daemonset-flow-readiness-probe-all.
+	namespace := "daemonset-flow-readiness-probe-all"
 	maybeRegisterCleanup(t, namespace)
 
 	testWorkflow(
 		t,
 		namespace,
-		csaQuantityAnnotationsCpuOnlyDefault, // TODO(wt) temporarily changed from csaQuantityAnnotationsAllDefault.
+		csaQuantityAnnotationsAllDefault,
 		func(annotations csaQuantityAnnotations) (string, int) {
 			config := echoDaemonSetConfigStandardPostStartup(namespace, annotations)
 			config.removeStartupProbes()